requirements.txt
flask==1.1.1
flask-cors==3.0.8
minio==5.0.7
arrow==0.15.5
flask-restful==0.3.8
redis==3.4.1
diskcache==4.1.0
#opencv-python-headless==4.2.0.32
#tensorflow==1.14.0
#tensorflow-gpu==1.14.0
#Keras==2.3.1
pymilvus==0.2.9
Pillow==7.0.0
gunicorn==19.10.0
tqdm==4.43.0
pytest==5.3.5
scipy
scikit-learn
requests
flask-sqlalchemy==2.4.1
mysqlclient==1.4.6
opencv-python
scikit-image
inflection
numpy
paddlepaddle
grpcio==1.27.2
# paddlepaddle-gpu==1.7.1.post107
